add highlighting support in text editor
uses peg-markdown-highlight [1].

this is a somewhat weird way of integrating the highlighter. it sits
in the background and waits for about 1 sec then parses the text and
creates an intermediate representation of the parsed syntactical
elements. then a highlighting thread (event handler) comes around
and applies styles to the `QDocument` embedded inside the
`QPlainTextEdit`.

[1]:http://hasseg.org/peg-markdown-highlight/